+++ cli/args/mod.rs
@@ -10,10 +10,6 @@ mod package_json;
 use std::borrow::Cow;
 use std::collections::HashMap;
 use std::env;
-use std::io::BufReader;
-use std::io::Cursor;
-use std::io::Read;
-use std::io::Seek;
 use std::net::SocketAddr;
 use std::num::NonZeroUsize;
 use std::path::Path;
@@ -85,8 +81,6 @@ pub use lockfile::CliLockfileReadFromPathOptions;
 use once_cell::sync::Lazy;
 pub use package_json::NpmInstallDepsProvider;
 pub use package_json::PackageJsonDepValueParseWithLocationError;
-use serde::Deserialize;
-use serde::Serialize;
 use sys_traits::EnvHomeDir;
 use thiserror::Error;
