+++ cli/build.rs
@@ -5,7 +5,6 @@ use std::path::PathBuf;
 
 use deno_core::snapshot::*;
 use deno_runtime::*;
-mod shared;
 
 mod ts {
   use std::collections::HashMap;
