+++ cli/rt/code_cache.rs
@@ -390,8 +390,6 @@ fn deserialize_with_reader<T: Read>(
 
 #[cfg(test)]
 mod test {
-  use std::fs::File;
-
   use test_util::TempDir;
 
   use super::*;
