+++ cli/rt/run.rs
@@ -1,3 +1,5 @@
+// Copyright 2018-2025 the Deno authors. MIT license.
+
 use std::borrow::Cow;
 use std::path::PathBuf;
 use std::rc::Rc;
