--- cli/shared.rs
@@ -1,1 +1,0 @@
-
