+++ cli/standalone/virtual_fs.rs
@@ -1,497 +1,21 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
 use std::borrow::Cow;
-use std::cell::RefCell;
-use std::cmp::Ordering;
-use std::collections::HashMap;
 use std::collections::HashSet;
-use std::fs::File;
-use std::io::Read;
-use std::io::Seek;
-use std::io::SeekFrom;
-use std::ops::Range;
-use std::path::Path;
 use std::path::PathBuf;
-use std::rc::Rc;
-use std::sync::Arc;
 
-use deno_core::anyhow::anyhow;
-use deno_core::anyhow::bail;
-use deno_core::anyhow::Context;
-use deno_core::error::AnyError;
-use deno_core::parking_lot::Mutex;
-use deno_core::BufMutView;
-use deno_core::BufView;
-use deno_core::ResourceHandleFd;
-use deno_lib::standalone::virtual_fs::FileSystemCaseSensitivity;
+use deno_lib::standalone::virtual_fs::BuiltVfs;
 use deno_lib::standalone::virtual_fs::OffsetWithLength;
 use deno_lib::standalone::virtual_fs::VfsEntry;
-use deno_lib::standalone::virtual_fs::VfsEntryRef;
-use deno_lib::standalone::virtual_fs::VfsFileSubDataKind;
 use deno_lib::standalone::virtual_fs::VirtualDirectory;
 use deno_lib::standalone::virtual_fs::VirtualDirectoryEntries;
 use deno_lib::standalone::virtual_fs::VirtualFile;
-use deno_lib::standalone::virtual_fs::VirtualSymlink;
 use deno_lib::standalone::virtual_fs::VirtualSymlinkParts;
 use deno_lib::standalone::virtual_fs::WindowsSystemRootablePath;
-use deno_path_util::normalize_path;
-use deno_path_util::strip_unc_prefix;
-use deno_runtime::deno_fs::FsDirEntry;
-use deno_runtime::deno_io;
-use deno_runtime::deno_io::fs::FsError;
-use deno_runtime::deno_io::fs::FsResult;
-use deno_runtime::deno_io::fs::FsStat;
-use indexmap::IndexSet;
-use serde::Deserialize;
-use serde::Serialize;
-use thiserror::Error;
+use deno_lib::standalone::virtual_fs::DENO_COMPILE_GLOBAL_NODE_MODULES_DIR_NAME;
 
-use super::binary::DENO_COMPILE_GLOBAL_NODE_MODULES_DIR_NAME;
-use crate::util;
 use crate::util::display::human_size;
 use crate::util::display::DisplayTreeNode;
-use crate::util::fs::canonicalize_path;
-
-#[derive(Debug)]
-pub struct BuiltVfs {
-  pub root_path: WindowsSystemRootablePath,
-  pub case_sensitivity: FileSystemCaseSensitivity,
-  pub entries: VirtualDirectoryEntries,
-  pub files: Vec<Vec<u8>>,
-}
-
-#[derive(Debug)]
-pub struct VfsBuilder {
-  executable_root: VirtualDirectory,
-  files: Vec<Vec<u8>>,
-  current_offset: u64,
-  file_offsets: HashMap<String, u64>,
-  /// The minimum root directory that should be included in the VFS.
-  min_root_dir: Option<WindowsSystemRootablePath>,
-  case_sensitivity: FileSystemCaseSensitivity,
-}
-
-impl VfsBuilder {
-  pub fn new() -> Self {
-    Self {
-      executable_root: VirtualDirectory {
-        name: "/".to_string(),
-        entries: Default::default(),
-      },
-      files: Vec::new(),
-      current_offset: 0,
-      file_offsets: Default::default(),
-      min_root_dir: Default::default(),
-      // This is not exactly correct because file systems on these OSes
-      // may be case-sensitive or not based on the directory, but this
-      // is a good enough approximation and limitation. In the future,
-      // we may want to store this information per directory instead
-      // depending on the feedback we get.
-      case_sensitivity: if cfg!(windows) || cfg!(target_os = "macos") {
-        FileSystemCaseSensitivity::Insensitive
-      } else {
-        FileSystemCaseSensitivity::Sensitive
-      },
-    }
-  }
-
-  pub fn case_sensitivity(&self) -> FileSystemCaseSensitivity {
-    self.case_sensitivity
-  }
-
-  /// Add a directory that might be the minimum root directory
-  /// of the VFS.
-  ///
-  /// For example, say the user has a deno.json and specifies an
-  /// import map in a parent directory. The import map won't be
-  /// included in the VFS, but its base will meaning we need to
-  /// tell the VFS builder to include the base of the import map
-  /// by calling this method.
-  pub fn add_possible_min_root_dir(&mut self, path: &Path) {
-    self.add_dir_raw(path);
-
-    match &self.min_root_dir {
-      Some(WindowsSystemRootablePath::WindowSystemRoot) => {
-        // already the root dir
-      }
-      Some(WindowsSystemRootablePath::Path(current_path)) => {
-        let mut common_components = Vec::new();
-        for (a, b) in current_path.components().zip(path.components()) {
-          if a != b {
-            break;
-          }
-          common_components.push(a);
-        }
-        if common_components.is_empty() {
-          if cfg!(windows) {
-            self.min_root_dir =
-              Some(WindowsSystemRootablePath::WindowSystemRoot);
-          } else {
-            self.min_root_dir =
-              Some(WindowsSystemRootablePath::Path(PathBuf::from("/")));
-          }
-        } else {
-          self.min_root_dir = Some(WindowsSystemRootablePath::Path(
-            common_components.iter().collect(),
-          ));
-        }
-      }
-      None => {
-        self.min_root_dir =
-          Some(WindowsSystemRootablePath::Path(path.to_path_buf()));
-      }
-    }
-  }
-
-  pub fn add_dir_recursive(&mut self, path: &Path) -> Result<(), AnyError> {
-    let target_path = self.resolve_target_path(path)?;
-    self.add_dir_recursive_not_symlink(&target_path)
-  }
-
-  fn add_dir_recursive_not_symlink(
-    &mut self,
-    path: &Path,
-  ) -> Result<(), AnyError> {
-    self.add_dir_raw(path);
-    let read_dir = std::fs::read_dir(path)
-      .with_context(|| format!("Reading {}", path.display()))?;
-
-    let mut dir_entries =
-      read_dir.into_iter().collect::<Result<Vec<_>, _>>()?;
-    dir_entries.sort_by_cached_key(|entry| entry.file_name()); // determinism
-
-    for entry in dir_entries {
-      let file_type = entry.file_type()?;
-      let path = entry.path();
-
-      if file_type.is_dir() {
-        self.add_dir_recursive_not_symlink(&path)?;
-      } else if file_type.is_file() {
-        self.add_file_at_path_not_symlink(&path)?;
-      } else if file_type.is_symlink() {
-        match self.add_symlink(&path) {
-          Ok(target) => match target {
-            SymlinkTarget::File(target) => {
-              self.add_file_at_path_not_symlink(&target)?
-            }
-            SymlinkTarget::Dir(target) => {
-              self.add_dir_recursive_not_symlink(&target)?;
-            }
-          },
-          Err(err) => {
-            log::warn!(
-            "{} Failed resolving symlink. Ignoring.\n    Path: {}\n    Message: {:#}",
-            crate::colors::yellow("Warning"),
-            path.display(),
-            err
-          );
-          }
-        }
-      }
-    }
-
-    Ok(())
-  }
-
-  fn add_dir_raw(&mut self, path: &Path) -> &mut VirtualDirectory {
-    log::debug!("Ensuring directory '{}'", path.display());
-    debug_assert!(path.is_absolute());
-    let mut current_dir = &mut self.executable_root;
-
-    for component in path.components() {
-      if matches!(component, std::path::Component::RootDir) {
-        continue;
-      }
-      let name = component.as_os_str().to_string_lossy();
-      let index = current_dir.entries.insert_or_modify(
-        &name,
-        self.case_sensitivity,
-        || {
-          VfsEntry::Dir(VirtualDirectory {
-            name: name.to_string(),
-            entries: Default::default(),
-          })
-        },
-        |_| {
-          // ignore
-        },
-      );
-      match current_dir.entries.get_mut_by_index(index) {
-        Some(VfsEntry::Dir(dir)) => {
-          current_dir = dir;
-        }
-        _ => unreachable!(),
-      };
-    }
-
-    current_dir
-  }
-
-  pub fn get_system_root_dir_mut(&mut self) -> &mut VirtualDirectory {
-    &mut self.executable_root
-  }
-
-  pub fn get_dir_mut(&mut self, path: &Path) -> Option<&mut VirtualDirectory> {
-    debug_assert!(path.is_absolute());
-    let mut current_dir = &mut self.executable_root;
-
-    for component in path.components() {
-      if matches!(component, std::path::Component::RootDir) {
-        continue;
-      }
-      let name = component.as_os_str().to_string_lossy();
-      let entry = current_dir
-        .entries
-        .get_mut_by_name(&name, self.case_sensitivity)?;
-      match entry {
-        VfsEntry::Dir(dir) => {
-          current_dir = dir;
-        }
-        _ => unreachable!(),
-      };
-    }
-
-    Some(current_dir)
-  }
-
-  pub fn add_file_at_path(&mut self, path: &Path) -> Result<(), AnyError> {
-    let file_bytes = std::fs::read(path)
-      .with_context(|| format!("Reading {}", path.display()))?;
-    self.add_file_with_data(path, file_bytes, VfsFileSubDataKind::Raw)
-  }
-
-  fn add_file_at_path_not_symlink(
-    &mut self,
-    path: &Path,
-  ) -> Result<(), AnyError> {
-    let file_bytes = std::fs::read(path)
-      .with_context(|| format!("Reading {}", path.display()))?;
-    self.add_file_with_data_inner(path, file_bytes, VfsFileSubDataKind::Raw)
-  }
-
-  pub fn add_file_with_data(
-    &mut self,
-    path: &Path,
-    data: Vec<u8>,
-    sub_data_kind: VfsFileSubDataKind,
-  ) -> Result<(), AnyError> {
-    let metadata = std::fs::symlink_metadata(path).with_context(|| {
-      format!("Resolving target path for '{}'", path.display())
-    })?;
-    if metadata.is_symlink() {
-      let target = self.add_symlink(path)?.into_path_buf();
-      self.add_file_with_data_inner(&target, data, sub_data_kind)
-    } else {
-      self.add_file_with_data_inner(path, data, sub_data_kind)
-    }
-  }
-
-  fn add_file_with_data_inner(
-    &mut self,
-    path: &Path,
-    data: Vec<u8>,
-    sub_data_kind: VfsFileSubDataKind,
-  ) -> Result<(), AnyError> {
-    log::debug!("Adding file '{}'", path.display());
-    let checksum = deno_lib::util::checksum::gen(&[&data]);
-    let case_sensitivity = self.case_sensitivity;
-    let offset = if let Some(offset) = self.file_offsets.get(&checksum) {
-      // duplicate file, reuse an old offset
-      *offset
-    } else {
-      self.file_offsets.insert(checksum, self.current_offset);
-      self.current_offset
-    };
-
-    let dir = self.add_dir_raw(path.parent().unwrap());
-    let name = path.file_name().unwrap().to_string_lossy();
-    let offset_and_len = OffsetWithLength {
-      offset,
-      len: data.len() as u64,
-    };
-    dir.entries.insert_or_modify(
-      &name,
-      case_sensitivity,
-      || {
-        VfsEntry::File(VirtualFile {
-          name: name.to_string(),
-          offset: offset_and_len,
-          module_graph_offset: offset_and_len,
-        })
-      },
-      |entry| match entry {
-        VfsEntry::File(virtual_file) => match sub_data_kind {
-          VfsFileSubDataKind::Raw => {
-            virtual_file.offset = offset_and_len;
-          }
-          VfsFileSubDataKind::ModuleGraph => {
-            virtual_file.module_graph_offset = offset_and_len;
-          }
-        },
-        VfsEntry::Dir(_) | VfsEntry::Symlink(_) => unreachable!(),
-      },
-    );
-
-    // new file, update the list of files
-    if self.current_offset == offset {
-      self.files.push(data);
-      self.current_offset += offset_and_len.len;
-    }
-
-    Ok(())
-  }
-
-  fn resolve_target_path(&mut self, path: &Path) -> Result<PathBuf, AnyError> {
-    let metadata = std::fs::symlink_metadata(path).with_context(|| {
-      format!("Resolving target path for '{}'", path.display())
-    })?;
-    if metadata.is_symlink() {
-      Ok(self.add_symlink(path)?.into_path_buf())
-    } else {
-      Ok(path.to_path_buf())
-    }
-  }
-
-  fn add_symlink(&mut self, path: &Path) -> Result<SymlinkTarget, AnyError> {
-    self.add_symlink_inner(path, &mut IndexSet::new())
-  }
-
-  fn add_symlink_inner(
-    &mut self,
-    path: &Path,
-    visited: &mut IndexSet<PathBuf>,
-  ) -> Result<SymlinkTarget, AnyError> {
-    log::debug!("Adding symlink '{}'", path.display());
-    let target = strip_unc_prefix(
-      std::fs::read_link(path)
-        .with_context(|| format!("Reading symlink '{}'", path.display()))?,
-    );
-    let case_sensitivity = self.case_sensitivity;
-    let target = normalize_path(path.parent().unwrap().join(&target));
-    let dir = self.add_dir_raw(path.parent().unwrap());
-    let name = path.file_name().unwrap().to_string_lossy();
-    dir.entries.insert_or_modify(
-      &name,
-      case_sensitivity,
-      || {
-        VfsEntry::Symlink(VirtualSymlink {
-          name: name.to_string(),
-          dest_parts: VirtualSymlinkParts::from_path(&target),
-        })
-      },
-      |_| {
-        // ignore previously inserted
-      },
-    );
-    let target_metadata =
-      std::fs::symlink_metadata(&target).with_context(|| {
-        format!("Reading symlink target '{}'", target.display())
-      })?;
-    if target_metadata.is_symlink() {
-      if !visited.insert(target.clone()) {
-        // todo: probably don't error in this scenario
-        bail!(
-          "Circular symlink detected: {} -> {}",
-          visited
-            .iter()
-            .map(|p| p.display().to_string())
-            .collect::<Vec<_>>()
-            .join(" -> "),
-          target.display()
-        );
-      }
-      self.add_symlink_inner(&target, visited)
-    } else if target_metadata.is_dir() {
-      Ok(SymlinkTarget::Dir(target))
-    } else {
-      Ok(SymlinkTarget::File(target))
-    }
-  }
-
-  pub fn build(self) -> BuiltVfs {
-    fn strip_prefix_from_symlinks(
-      dir: &mut VirtualDirectory,
-      parts: &[String],
-    ) {
-      for entry in dir.entries.iter_mut() {
-        match entry {
-          VfsEntry::Dir(dir) => {
-            strip_prefix_from_symlinks(dir, parts);
-          }
-          VfsEntry::File(_) => {}
-          VfsEntry::Symlink(symlink) => {
-            let parts = symlink
-              .dest_parts
-              .take_parts()
-              .into_iter()
-              .skip(parts.len())
-              .collect();
-            symlink.dest_parts.set_parts(parts);
-          }
-        }
-      }
-    }
-
-    let mut current_dir = self.executable_root;
-    let mut current_path = if cfg!(windows) {
-      WindowsSystemRootablePath::WindowSystemRoot
-    } else {
-      WindowsSystemRootablePath::Path(PathBuf::from("/"))
-    };
-    loop {
-      if current_dir.entries.len() != 1 {
-        break;
-      }
-      if self.min_root_dir.as_ref() == Some(&current_path) {
-        break;
-      }
-      match current_dir.entries.iter().next().unwrap() {
-        VfsEntry::Dir(dir) => {
-          if dir.name == DENO_COMPILE_GLOBAL_NODE_MODULES_DIR_NAME {
-            // special directory we want to maintain
-            break;
-          }
-          match current_dir.entries.remove(0) {
-            VfsEntry::Dir(dir) => {
-              current_path =
-                WindowsSystemRootablePath::Path(current_path.join(&dir.name));
-              current_dir = dir;
-            }
-            _ => unreachable!(),
-          };
-        }
-        VfsEntry::File(_) | VfsEntry::Symlink(_) => break,
-      }
-    }
-    if let WindowsSystemRootablePath::Path(path) = &current_path {
-      strip_prefix_from_symlinks(
-        &mut current_dir,
-        VirtualSymlinkParts::from_path(path).parts(),
-      );
-    }
-    BuiltVfs {
-      root_path: current_path,
-      case_sensitivity: self.case_sensitivity,
-      entries: current_dir.entries,
-      files: self.files,
-    }
-  }
-}
-
-#[derive(Debug)]
-enum SymlinkTarget {
-  File(PathBuf),
-  Dir(PathBuf),
-}
-
-impl SymlinkTarget {
-  pub fn into_path_buf(self) -> PathBuf {
-    match self {
-      Self::File(path) => path,
-      Self::Dir(path) => path,
-    }
-  }
-}
 
 pub fn output_vfs(vfs: &BuiltVfs, executable_name: &str) {
   if !log::log_enabled!(log::Level::Info) {
@@ -843,281 +367,11 @@ fn vfs_as_display_tree(
 
 #[cfg(test)]
 mod test {
-  use std::io::Write;
-
   use console_static_text::ansi::strip_ansi_codes;
-  use deno_io::fs::File;
-  use test_util::assert_contains;
+  use deno_lib::standalone::virtual_fs::VfsBuilder;
   use test_util::TempDir;
 
   use super::*;
-
-  #[track_caller]
-  fn read_file(vfs: &FileBackedVfs, path: &Path) -> String {
-    let file = vfs.file_entry(path).unwrap();
-    String::from_utf8(
-      vfs
-        .read_file_all(file, VfsFileSubDataKind::Raw)
-        .unwrap()
-        .into_owned(),
-    )
-    .unwrap()
-  }
-
-  #[test]
-  fn builds_and_uses_virtual_fs() {
-    let temp_dir = TempDir::new();
-    // we canonicalize the temp directory because the vfs builder
-    // will canonicalize the root path
-    let src_path = temp_dir.path().canonicalize().join("src");
-    src_path.create_dir_all();
-    src_path.join("sub_dir").create_dir_all();
-    src_path.join("e.txt").write("e");
-    src_path.symlink_file("e.txt", "sub_dir/e.txt");
-    let src_path = src_path.to_path_buf();
-    let mut builder = VfsBuilder::new();
-    builder
-      .add_file_with_data_inner(
-        &src_path.join("a.txt"),
-        "data".into(),
-        VfsFileSubDataKind::Raw,
-      )
-      .unwrap();
-    builder
-      .add_file_with_data_inner(
-        &src_path.join("b.txt"),
-        "data".into(),
-        VfsFileSubDataKind::Raw,
-      )
-      .unwrap();
-    assert_eq!(builder.files.len(), 1); // because duplicate data
-    builder
-      .add_file_with_data_inner(
-        &src_path.join("c.txt"),
-        "c".into(),
-        VfsFileSubDataKind::Raw,
-      )
-      .unwrap();
-    builder
-      .add_file_with_data_inner(
-        &src_path.join("sub_dir").join("d.txt"),
-        "d".into(),
-        VfsFileSubDataKind::Raw,
-      )
-      .unwrap();
-    builder.add_file_at_path(&src_path.join("e.txt")).unwrap();
-    builder
-      .add_symlink(&src_path.join("sub_dir").join("e.txt"))
-      .unwrap();
-
-    // get the virtual fs
-    let (dest_path, virtual_fs) = into_virtual_fs(builder, &temp_dir);
-
-    assert_eq!(read_file(&virtual_fs, &dest_path.join("a.txt")), "data");
-    assert_eq!(read_file(&virtual_fs, &dest_path.join("b.txt")), "data");
-
-    // attempt reading a symlink
-    assert_eq!(
-      read_file(&virtual_fs, &dest_path.join("sub_dir").join("e.txt")),
-      "e",
-    );
-
-    // canonicalize symlink
-    assert_eq!(
-      virtual_fs
-        .canonicalize(&dest_path.join("sub_dir").join("e.txt"))
-        .unwrap(),
-      dest_path.join("e.txt"),
-    );
-
-    // metadata
-    assert_eq!(
-      virtual_fs
-        .lstat(&dest_path.join("sub_dir").join("e.txt"))
-        .unwrap()
-        .file_type,
-      sys_traits::FileType::Symlink,
-    );
-    assert_eq!(
-      virtual_fs
-        .stat(&dest_path.join("sub_dir").join("e.txt"))
-        .unwrap()
-        .file_type,
-      sys_traits::FileType::File,
-    );
-    assert_eq!(
-      virtual_fs
-        .stat(&dest_path.join("sub_dir"))
-        .unwrap()
-        .file_type,
-      sys_traits::FileType::Dir,
-    );
-    assert_eq!(
-      virtual_fs.stat(&dest_path.join("e.txt")).unwrap().file_type,
-      sys_traits::FileType::File
-    );
-  }
-
-  #[test]
-  fn test_include_dir_recursive() {
-    let temp_dir = TempDir::new();
-    let temp_dir_path = temp_dir.path().canonicalize();
-    temp_dir.create_dir_all("src/nested/sub_dir");
-    temp_dir.write("src/a.txt", "data");
-    temp_dir.write("src/b.txt", "data");
-    util::fs::symlink_dir(
-      &crate::sys::CliSys::default(),
-      temp_dir_path.join("src/nested/sub_dir").as_path(),
-      temp_dir_path.join("src/sub_dir_link").as_path(),
-    )
-    .unwrap();
-    temp_dir.write("src/nested/sub_dir/c.txt", "c");
-
-    // build and create the virtual fs
-    let src_path = temp_dir_path.join("src").to_path_buf();
-    let mut builder = VfsBuilder::new();
-    builder.add_dir_recursive(&src_path).unwrap();
-    let (dest_path, virtual_fs) = into_virtual_fs(builder, &temp_dir);
-
-    assert_eq!(read_file(&virtual_fs, &dest_path.join("a.txt")), "data",);
-    assert_eq!(read_file(&virtual_fs, &dest_path.join("b.txt")), "data",);
-
-    assert_eq!(
-      read_file(
-        &virtual_fs,
-        &dest_path.join("nested").join("sub_dir").join("c.txt")
-      ),
-      "c",
-    );
-    assert_eq!(
-      read_file(&virtual_fs, &dest_path.join("sub_dir_link").join("c.txt")),
-      "c",
-    );
-    assert_eq!(
-      virtual_fs
-        .lstat(&dest_path.join("sub_dir_link"))
-        .unwrap()
-        .file_type,
-      sys_traits::FileType::Symlink,
-    );
-
-    assert_eq!(
-      virtual_fs
-        .canonicalize(&dest_path.join("sub_dir_link").join("c.txt"))
-        .unwrap(),
-      dest_path.join("nested").join("sub_dir").join("c.txt"),
-    );
-  }
-
-  fn into_virtual_fs(
-    builder: VfsBuilder,
-    temp_dir: &TempDir,
-  ) -> (PathBuf, FileBackedVfs) {
-    let virtual_fs_file = temp_dir.path().join("virtual_fs");
-    let vfs = builder.build();
-    {
-      let mut file = std::fs::File::create(&virtual_fs_file).unwrap();
-      for file_data in &vfs.files {
-        file.write_all(file_data).unwrap();
-      }
-    }
-    let dest_path = temp_dir.path().join("dest");
-    let data = std::fs::read(&virtual_fs_file).unwrap();
-    (
-      dest_path.to_path_buf(),
-      FileBackedVfs::new(
-        Cow::Owned(data),
-        VfsRoot {
-          dir: VirtualDirectory {
-            name: "".to_string(),
-            entries: vfs.entries,
-          },
-          root_path: dest_path.to_path_buf(),
-          start_file_offset: 0,
-        },
-        FileSystemCaseSensitivity::Sensitive,
-      ),
-    )
-  }
-
-  #[test]
-  fn circular_symlink() {
-    let temp_dir = TempDir::new();
-    let src_path = temp_dir.path().canonicalize().join("src");
-    src_path.create_dir_all();
-    src_path.symlink_file("a.txt", "b.txt");
-    src_path.symlink_file("b.txt", "c.txt");
-    src_path.symlink_file("c.txt", "a.txt");
-    let src_path = src_path.to_path_buf();
-    let mut builder = VfsBuilder::new();
-    let err = builder
-      .add_symlink(src_path.join("a.txt").as_path())
-      .unwrap_err();
-    assert_contains!(err.to_string(), "Circular symlink detected",);
-  }
-
-  #[tokio::test]
-  async fn test_open_file() {
-    let temp_dir = TempDir::new();
-    let temp_path = temp_dir.path().canonicalize();
-    let mut builder = VfsBuilder::new();
-    builder
-      .add_file_with_data_inner(
-        temp_path.join("a.txt").as_path(),
-        "0123456789".to_string().into_bytes(),
-        VfsFileSubDataKind::Raw,
-      )
-      .unwrap();
-    let (dest_path, virtual_fs) = into_virtual_fs(builder, &temp_dir);
-    let virtual_fs = Arc::new(virtual_fs);
-    let file = virtual_fs.open_file(&dest_path.join("a.txt")).unwrap();
-    file.seek(SeekFrom::Current(2)).unwrap();
-    let mut buf = vec![0; 2];
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"23");
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"45");
-    file.seek(SeekFrom::Current(-4)).unwrap();
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"23");
-    file.seek(SeekFrom::Start(2)).unwrap();
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"23");
-    file.seek(SeekFrom::End(2)).unwrap();
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"89");
-    file.seek(SeekFrom::Current(-8)).unwrap();
-    file.read_to_buf(&mut buf).unwrap();
-    assert_eq!(buf, b"23");
-    assert_eq!(
-      file
-        .seek(SeekFrom::Current(-5))
-        .unwrap_err()
-        .to_string(),
-      "An attempt was made to move the file pointer before the beginning of the file."
-    );
-    // go beyond the file length, then back
-    file.seek(SeekFrom::Current(40)).unwrap();
-    file.seek(SeekFrom::Current(-38)).unwrap();
-    let file = Rc::new(file);
-    let read_buf = file.clone().read(2).await.unwrap();
-    assert_eq!(read_buf.to_vec(), b"67");
-    file.clone().seek_sync(SeekFrom::Current(-2)).unwrap();
-
-    // read to the end of the file
-    let all_buf = file.clone().read_all_sync().unwrap();
-    assert_eq!(all_buf.to_vec(), b"6789");
-    file.clone().seek_sync(SeekFrom::Current(-9)).unwrap();
-
-    // try try_clone_inner and read_all_async
-    let all_buf = file
-      .try_clone_inner()
-      .unwrap()
-      .read_all_async()
-      .await
-      .unwrap();
-    assert_eq!(all_buf.to_vec(), b"123456789");
-  }
 
   #[test]
   fn test_vfs_as_display_tree() {
