+++ cli/util/text_encoding.rs
@@ -1,8 +1,6 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
-use std::borrow::Cow;
 use std::ops::Range;
-use std::sync::Arc;
 
 use base64::prelude::BASE64_STANDARD;
 use base64::Engine;
